+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ### Fixed
 - Fixed nullable array [#1094](https://github.com/zephir-lang/zephir/issues/1094)
 
+### Added
+- Added support syntax assign-bitwise operators [#1103](https://github.com/zephir-lang/zephir/issues/1103)
+
 ## [0.13.2] - 2021-04-10
 ### Fixed
 - Fixed default value of nullable string parameter [#2180](https://github.com/zephir-lang/zephir/issues/2180)