+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
+
+
+## [0.13.3] - 2021-04-24
 ### Fixed
 - Fixed nullable array [#1094](https://github.com/zephir-lang/zephir/issues/1094)
 - Fixed default value detection with Reflection (only PHP 8.0) [#1134](https://github.com/zephir-lang/zephir/issues/1134)
@@ -462,8 +465,9 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - Fixed casting resource to int (only ZendEngine 3)
   [#1524](https://github.com/zephir-lang/zephir/issues/1524)
 
-[Unreleased]: https://github.com/zephir-lang/zephir/compare/0.13.2...HEAD
-[0.13.2]: https://github.com/zephir-lang/zephir/compare/0.13.0...0.13.2
+[Unreleased]: https://github.com/zephir-lang/zephir/compare/0.13.3...HEAD
+[0.13.3]: https://github.com/zephir-lang/zephir/compare/0.13.2...0.13.3
+[0.13.2]: https://github.com/zephir-lang/zephir/compare/0.13.1...0.13.2
 [0.13.1]: https://github.com/zephir-lang/zephir/compare/0.13.0...0.13.1
 [0.13.0]: https://github.com/zephir-lang/zephir/compare/0.12.21...0.13.0
 [0.12.21]: https://github.com/zephir-lang/zephir/compare/0.12.20...0.12.21